pyproject.toml
[tool.poetry]
name = "file-uploader"
version = "0.1.0"
description = ""
authors = ["Sergey Rubnikovich <sergeyrubnikovich@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.87.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
python-multipart = "^0.0.5"
orjson = "^3.8.2"
SQLAlchemy = "^1.4.44"
asyncpg = "^0.27.0"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
aiofiles = "^22.1.0"
httpx = "^0.24.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
flake8 = "^6.0.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
